Feat: Enables navigation to pilot details
Enables navigation to the pilot details page from the race results list.

This allows users to quickly access comprehensive information about
a specific pilot by tapping on their name or image in the race results.

It creates a minimal event object from the round data, and passes it to the
PilotDetails component.